requirements.txt
keyboa==3.1.0
psycopg2==2.9.3
pyTelegramBotAPI==4.6.0
pyyaml=6.0
